Split the utilities into seperate files
This allows users to only include the specific utilities that
they need.
