+++ src/CMakeLists.txt
@@ -7,5 +7,7 @@ set(${PROJECT_NAME}_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/mutual_info.c
     ${CMAKE_CURRENT_SOURCE_DIR}/shannon.c
     ${CMAKE_CURRENT_SOURCE_DIR}/transfer_entropy.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/utilities.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/utilities/binning.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/utilities/coalesce.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/utilities/encoding.c
     PARENT_SCOPE)
